LPS-193776 Move ErrorAlert to layout_reports folder as it's used there
