+++ modules/apps/layout/layout-reports-web/src/main/resources/META-INF/resources/js/components/layout_reports/ErrorAlertExtendedInfo.js
