+++ modules/apps/layout/layout-reports-web/src/main/resources/META-INF/resources/js/components/layout_reports/LayoutReports.js
@@ -18,8 +18,8 @@ import {
 	StoreStateContext,
 } from '../../context/StoreContext';
 import loadIssues from '../../utils/loadIssues';
-import ErrorAlert from '../error_alert/ErrorAlert';
 import BasicInformation from './BasicInformation';
+import ErrorAlert from './ErrorAlert';
 import IssueDetail from './IssueDetail';
 import IssuesList from './IssuesList';
 import NotConfigured from './NotConfigured';
